Refactor pipelinerun matchings to its own
Refactor matchRepoByPr function to its own file
Refactor getAllPipelineRuns to the PacRun struct and move that to the
match file too.
